+++ out/configs.js
@@ -1,0 +1,72 @@
+"use strict";
+var __createBinding = (this && this.__createBinding) || (Object.create ? (function(o, m, k, k2) {
+    if (k2 === undefined) k2 = k;
+    var desc = Object.getOwnPropertyDescriptor(m, k);
+    if (!desc || ("get" in desc ? !m.__esModule : desc.writable || desc.configurable)) {
+      desc = { enumerable: true, get: function() { return m[k]; } };
+    }
+    Object.defineProperty(o, k2, desc);
+}) : (function(o, m, k, k2) {
+    if (k2 === undefined) k2 = k;
+    o[k2] = m[k];
+}));
+var __setModuleDefault = (this && this.__setModuleDefault) || (Object.create ? (function(o, v) {
+    Object.defineProperty(o, "default", { enumerable: true, value: v });
+}) : function(o, v) {
+    o["default"] = v;
+});
+var __importStar = (this && this.__importStar) || function (mod) {
+    if (mod && mod.__esModule) return mod;
+    var result = {};
+    if (mod != null) for (var k in mod) if (k !== "default" && Object.prototype.hasOwnProperty.call(mod, k)) __createBinding(result, mod, k);
+    __setModuleDefault(result, mod);
+    return result;
+};
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.Configs = void 0;
+const vscode = __importStar(require("vscode"));
+class Configs {
+    static enabledCursor = false;
+    static enabledExtension = false;
+    static enabledGrayscale = false;
+    static get isCursorEnabled() { return this.enabledCursor; }
+    static get isExtensionEnabled() { return this.enabledExtension; }
+    static get isGrayscaleEnabled() { return this.enabledGrayscale; }
+    constructor() { }
+    static activate(context) {
+        const extName = 'funnycode';
+        const configs = new Map([
+            ['enabled', (v) => Configs.enabledExtension = v],
+            ['cursor', (v) => Configs.enabledCursor = v],
+            ['grayscale', (v) => Configs.enabledGrayscale = v],
+        ]);
+        function isConfigEnabled(config) {
+            const cfg = vscode.workspace.getConfiguration(extName);
+            const inspect = cfg.inspect(config);
+            return !!(inspect?.globalValue ?? inspect?.defaultValue ?? false);
+        }
+        function toggleExtensionEnabled() {
+            const configName = 'enabled';
+            const value = isConfigEnabled(configName);
+            const config = vscode.workspace.getConfiguration(extName);
+            config.update(configName, !value, vscode.ConfigurationTarget.Global);
+        }
+        function onConfigChanged(event) {
+            for (const [name, setter] of configs) {
+                if (event.affectsConfiguration(`${extName}.${name}`)) {
+                    setter(isConfigEnabled(name));
+                }
+            }
+        }
+        for (const [name, setter] of configs) {
+            setter(isConfigEnabled(name));
+        }
+        const toggleCommandId = 'funnycode.toggleEnable';
+        const toogleCommand = vscode.commands.registerCommand(toggleCommandId, toggleExtensionEnabled);
+        context.subscriptions.push(toogleCommand);
+        const configSub = vscode.workspace.onDidChangeConfiguration(onConfigChanged);
+        context.subscriptions.push(configSub);
+    }
+}
+exports.Configs = Configs;
+//# sourceMappingURL=configs.js.map